+++ dsb-gui/src/main/java/com/spectralogic/dsbrowser/gui/services/jobService/factories/FactoryExtensions.kt
@@ -33,46 +33,3 @@ import java.util.*
 fun removeJob(uuid: UUID, client: Ds3Client, jobInterruptionStore: JobInterruptionStore, deepStorageBrowserPresenter: DeepStorageBrowserPresenter, loggingService: LoggingService) {
     ParseJobInterruptionMap.removeJobID(jobInterruptionStore, uuid.toString(), client.getConnectionDetails().getEndpoint(), deepStorageBrowserPresenter, loggingService)
 }
-
-fun JobTask.onSucceeded(type: String, log: Logger): (WorkerStateEvent) -> Unit = {
-    log.info("$type Job completed successfully")
-}
-
-fun JobTask.onFailed(client: Ds3Client,
-                     jobInterruptionStore: JobInterruptionStore,
-                     deepStorageBrowserPresenter: DeepStorageBrowserPresenter,
-                     loggingService: LoggingService,
-                     log: Logger,
-                     type: String): (WorkerStateEvent) -> Unit = { worker: WorkerStateEvent ->
-    val throwable: Throwable = worker.source.exception
-    jobId.exists {
-        removeJob(it, client, jobInterruptionStore, deepStorageBrowserPresenter, loggingService)
-    }
-    log.error("$type Job failed", throwable)
-    loggingService.logMessage("$type Job failed with message: ${throwable.message}", LogType.ERROR)
-}
-
-fun JobTask.onCancelled(client: Ds3Client,
-                        type: String,
-                        log: Logger,
-                        loggingService: LoggingService,
-                        jobInterruptionStore: JobInterruptionStore,
-                        deepStorageBrowserPresenter: DeepStorageBrowserPresenter): (WorkerStateEvent) -> Unit = {
-    jobId.exists {
-        try {
-            Platform.runLater {
-                client.cancelJobSpectraS3(CancelJobSpectraS3Request(it))
-            }
-        } catch (e: IOException) {
-            log.error("Failed to cancel $type job", e)
-            loggingService.logMessage("Could not cancel $type job", LogType.ERROR)
-        }
-        log.info("$type Job cancelled")
-        loggingService.logMessage("$type Job Cancelled", LogType.INFO)
-        removeJob(it, client, jobInterruptionStore, deepStorageBrowserPresenter, loggingService)
-    }
-
-}
-
-fun JobTask.onRunning() = { _: WorkerStateEvent -> }
-fun JobTask.onScheduled() = { _: WorkerStateEvent -> }
